requirements.txt
streamlit
requests
tqdm
pandas
numpy
plotly
gcsfs
st-files-connection
google-cloud-run